ref: user情報の取得をhooksに寄せる
